+++ plugins/jobs/drivers/ephemeral/item.go
@@ -111,6 +111,6 @@ func (i *Item) Nack() error {
 	return nil
 }
 
-func (i *Item) Requeue(_ uint32) error {
+func (i *Item) Requeue(_ int64) error {
 	return nil
 }
